+++ dmr/security/__init__.py
@@ -2,7 +2,7 @@ from dmr.security.base import NO_STORE_HEADERS as NO_STORE_HEADERS
 from dmr.security.base import AsyncAuth as AsyncAuth
 from dmr.security.base import SyncAuth as SyncAuth
 from dmr.security.base import SyncOrAsyncAuth as SyncOrAsyncAuth
-from dmr.security.base import add_auth_challenge as add_auth_challenge
+from dmr.security.base import add_www_authenticate as add_www_authenticate
 from dmr.security.base import request_auth as request_auth
 from dmr.security.types import (
     AuthenticatedHttpRequest as AuthenticatedHttpRequest,